modified age param to be passed as long containing date and check for matching date in age col of patients table

+++ app/femr/business/services/core/ISearchService.java
@@ -108,7 +108,7 @@ public interface ISearchService {
      */
     ServiceResponse<List<PatientItem>> retrievePatientsFromQueryString(String patientSearchQuery);
 
-    ServiceResponse<List<PatientItem>> retrievePatientsFromTriageSearch(String first, String last, String phone, String addr, String gender, int age, String city);
+    ServiceResponse<List<PatientItem>> retrievePatientsFromTriageSearch(String first, String last, String phone, String addr, String gender, Long age, String city);
 
 
     /**
